moved functions into module
