--- innovation_sweet_spots/analysis/query_lda_model.py
@@ -1,31 +1,0 @@
-"""
-innovation_sweet_spots.analysis.query_lda_model
-
-Module for selecting items using LDA topic model
-"""
-# from innovation_sweet_spots import logging
-# import numpy as np
-# import pandas as pd
-# from typing import Iterator, Dict, Union
-# from numpy.typing import ArrayLike
-
-# Model getter (maybe move into model)
-def get_pilot_lda_model():
-    """
-    mdl = tp.LDAModel.load(str(PROJECT_DIR / "outputs/models/full_corpus_lda_model.bin"))
-    Returns:
-        A dictionary with 'the model' and a table with topic definitions/keywords
-    """
-    pass
-
-
-def get_topic_probabilities():
-    """ """
-    pass
-
-
-def query_topics(topics: Iterator[int], topic_model) -> pd.DataFrame:
-    pass
-
-
-#
